starter/apps/order-api/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
httpx==0.28.1
boto3==1.36.14
pydantic==2.10.6
dnspython==2.7.0
python-json-logger==3.2.1
prometheus-client==0.20.0
